fixed a floating point conversion when using constrained ICs

+++ constraints.cc
@@ -170,7 +170,7 @@ constraint_set::constraint_set( config_file& cf, transfer_function *ptf )
 	
 	unsigned i=0;
 	
-    double astart = 1.0/(1.0+pcf_->getValue<unsigned>("setup","zstart"));
+    double astart = 1.0/(1.0+pcf_->getValue<double>("setup","zstart"));
     unsigned levelmin = pcf_->getValue<unsigned>("setup","levelmin");
 	unsigned levelmin_TF = pcf_->getValueSafe<unsigned>("setup","levelmin_TF",levelmin);
 	constr_level_ = pcf_->getValueSafe<unsigned>("constraints","level",levelmin_TF);
